requirements.txt
google-generativeai==0.3.2
Pillow==10.1.0